backend/face_recognition_app/services_simple.py
# Temporarily simplified for basic setup
import cv2
import numpy as np
from PIL import Image
import os
import time
from django.conf import settings
from django.contrib.auth import get_user_model
from .models import FaceEncoding, FaceRecognitionLog

User = get_user_model()


class FaceRecognitionService:
    """Service class for face recognition operations - simplified for setup"""
    
    def __init__(self):
        self.tolerance = getattr(settings, 'FACE_RECOGNITION_TOLERANCE', 0.6)
        self.model = getattr(settings, 'FACE_RECOGNITION_MODEL', 'hog')
    
    def get_face_locations(self, image_path_or_array):
        """Get face locations from image - placeholder"""
        # TODO: Implement when face_recognition library is installed
        return []
    
    def get_face_encodings(self, image_path_or_array, face_locations=None):
        """Get face encodings from image - placeholder"""
        # TODO: Implement when face_recognition library is installed
        return []
    
    def add_person(self, person_id, name, face_encoding):
        """Add a new person to the recognition database"""
        try:
            user = User.objects.get(id=person_id)
            
            # Save face encoding
            face_enc = FaceEncoding.objects.create(
                user=user,
                encoding=face_encoding.tobytes() if isinstance(face_encoding, np.ndarray) else face_encoding,
                is_active=True
            )
            
            return {"success": True, "message": f"Person {name} added successfully"}
        except Exception as e:
            return {"success": False, "message": f"Error adding person: {str(e)}"}
    
    def recognize_faces(self, image_path_or_array):
        """Recognize faces in the given image - placeholder"""
        try:
            # TODO: Implement face recognition when library is available
            return {
                "success": True,
                "faces": [],
                "message": "Face recognition temporarily disabled"
            }
        except Exception as e:
            return {
                "success": False,
                "faces": [],
                "message": f"Error in face recognition: {str(e)}"
            }
    
    def verify_face(self, known_encoding, test_encoding):
        """Verify if two face encodings match - placeholder"""
        try:
            # TODO: Implement when face_recognition library is available
            return {"match": False, "distance": 1.0}
        except Exception as e:
            return {"match": False, "distance": 1.0, "error": str(e)}
    
    def get_face_quality_score(self, image_path_or_array):
        """Analyze face quality in image - placeholder"""
        try:
            # TODO: Implement quality analysis
            return {
                "quality_score": 0.5,
                "issues": ["Face recognition library not available"],
                "recommendations": ["Install face_recognition library"]
            }
        except Exception as e:
            return {
                "quality_score": 0.0,
                "issues": [f"Error: {str(e)}"],
                "recommendations": ["Check image format and quality"]
            }
    
    def bulk_train(self, force_retrain=False):
        """Train the face recognition model with all available data - placeholder"""
        try:
            # TODO: Implement bulk training
            return {
                "success": True,
                "message": "Training temporarily disabled",
                "processed_faces": 0
            }
        except Exception as e:
            return {
                "success": False,
                "message": f"Training error: {str(e)}",
                "processed_faces": 0
            }