backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
gunicorn==21.2.0
pytest-django==4.7.0
pytest==7.4.3
factory-boy==3.3.0
coverage==7.3.2
channels==4.0.0
channels-redis==4.1.0
django-filter==23.5